requirements.txt
requests==2.32.3
python-dotenv==1.0.1
mysql-connector-python==9.2.0
